refactor: regroup src/components by domain, standardize tests, refresh README
Components moved out of a flat dump into domain folders:
- common/   CircularProgress, ProgressCard, Header, ErrorBoundary
- themed/   ThemedText, ThemedView
- charts/   MiniCharts, WeightBarChart, ActivityRings, ActivityHeatmap
- workout/  Workout, WorkoutDetail, WorkoutProgress, WorkoutCompleteModal, RestTimer, VideoPlayer
- dashboard/ CalendarStrip, WorkoutXPCard
- health/   DailySteps, HealthMetrics
- weight/   WeightGoalSheet

Tests:
- Standardize on *.test.ts(x) (was a mix of *-test.ts(x) and *.test.ts(x))
- Colocate component tests in domain-local __tests__/ folders
- Move orphan src/__tests__/WorkoutProgress.test.tsx next to its component
- Rename snapshot file to match the renamed test

lib <-> utils:
- Move pure recovery math from src/lib/recovery.ts to src/utils/recovery.ts
- Drop the 1-line re-export shim at src/utils/recoveryScore.ts
- Consumers now import from @/utils/recovery directly

README rewrite:
- Section 3 (Project Structure): reflects the real tree, all lib/* feature folders, all screens, test conventions
- Section 4 (Architecture): full store roster + offline-first/Zod/HealthKit-adapter flow
- Section 5 (Database Schema): documents all 5 tables (exercises, weight_entries, daily_health_snapshots, workout_sessions, meals)
- Section 6 (Data Validation): lists every Zod schema
- Section 8: fix stale plugin list (@kingstinct/react-native-healthkit)

Also includes pre-existing local edits to nutrition/scan.tsx and removal of src/lib/aiParser/index.ts (parseMeal now imported directly from mockParser); these were already in the working tree and pass typecheck/lint/tests.

Quality gate: bun run quality:pr -> expo:check, expo:doctor, lint, typecheck, 21 suites / 154 tests, 1 snapshot all green.
