+++ src/components/charts/WeightBarChart.tsx
