+++ src/components/dashboard/CalendarStrip.tsx
