+++ src/stores/__tests__/WeightStore.test.ts
