--- src/utils/recoveryScore.ts
@@ -1,1 +1,0 @@
-export * from '@/lib/recovery'
